feat: Complete RBAC system implementation with comprehensive permissions
- Add 61 comprehensive permissions for all modules (cate_product, product, cate_news, news, slider, menu, feedback, contact, comment, user, system, analytics, page, order, product_setting, profile, role, permission)
- Implement CheckPermission middleware for all admin routes
- Update navbar with proper @hasAnyPermission directives
- Fix blade directive registration in AppServiceProvider
- Create UserRoleTestSeeder with 4 test accounts
- Clean up debug routes

Test Accounts:
- admin@gmail.com (@admin123) - Full admin access (61 permissions)
- manager@gmail.com (@manager123) - Management access (40+ permissions)
- staff@gmail.com (@staff123) - Limited staff access (15+ permissions)
- editor@gmail.com (@editor123) - Content editor only (14 permissions)

+++ routes/admin/user-role.php
@@ -4,16 +4,16 @@ use App\Http\Controllers\Admin\UserRoleController;
 use Illuminate\Support\Facades\Route;
 // User Role Management Routes
 Route::prefix('user-role')->name('user-role.')->group(function () {
-    Route::get('/', [UserRoleController::class, 'index'])->name('index');
-    Route::get('/{id}/edit', [UserRoleController::class, 'edit'])->name('edit');
-    Route::put('/{id}', [UserRoleController::class, 'updateRoles'])->name('update');
-    Route::post('/bulk-update', [UserRoleController::class, 'bulkUpdateRoles'])->name('bulk-update');
+    Route::get('/', [UserRoleController::class, 'index'])->name('index')->middleware('permission:user.view');
+    Route::get('/{id}/edit', [UserRoleController::class, 'edit'])->name('edit')->middleware('permission:user.edit');
+    Route::put('/{id}', [UserRoleController::class, 'updateRoles'])->name('update')->middleware('permission:user.edit');
+    Route::post('/bulk-update', [UserRoleController::class, 'bulkUpdateRoles'])->name('bulk-update')->middleware('permission:user.edit');
     
     // Role management routes
-    Route::get('/roles', [UserRoleController::class, 'roleIndex'])->name('roles.index');
-    Route::get('/roles/create', [UserRoleController::class, 'roleCreate'])->name('roles.create');
-    Route::post('/roles', [UserRoleController::class, 'roleStore'])->name('roles.store');
-    Route::get('/roles/{id}/edit', [UserRoleController::class, 'roleEdit'])->name('roles.edit');
-    Route::put('/roles/{id}', [UserRoleController::class, 'roleUpdate'])->name('roles.update');
-    Route::delete('/roles/{id}', [UserRoleController::class, 'roleDestroy'])->name('roles.destroy');
+    Route::get('/roles', [UserRoleController::class, 'roleIndex'])->name('roles.index')->middleware('permission:user.view');
+    Route::get('/roles/create', [UserRoleController::class, 'roleCreate'])->name('roles.create')->middleware('permission:user.create');
+    Route::post('/roles', [UserRoleController::class, 'roleStore'])->name('roles.store')->middleware('permission:user.create');
+    Route::get('/roles/{id}/edit', [UserRoleController::class, 'roleEdit'])->name('roles.edit')->middleware('permission:user.edit');
+    Route::put('/roles/{id}', [UserRoleController::class, 'roleUpdate'])->name('roles.update')->middleware('permission:user.edit');
+    Route::delete('/roles/{id}', [UserRoleController::class, 'roleDestroy'])->name('roles.destroy')->middleware('permission:user.delete');
 });
